move interaction-state to canvas-strategies
