+++ editor/src/components/canvas/controls/size-box.tsx
@@ -46,7 +46,7 @@ import {
 import { safeIndex } from '../../../core/shared/array-utils'
 import { CSSPosition } from '../../inspector/common/css-utils'
 import * as EP from '../../../core/shared/element-path'
-import { createInteractionViaMouse } from '../interactions/interaction-state'
+import { createInteractionViaMouse } from '../canvas-strategies/interaction-state'
 import { Modifier } from '../../../utils/modifiers'
 import { windowToCanvasCoordinates } from '../dom-lookup'
