+++ editor/src/components/canvas/ui-jsx.test-utils.tsx
@@ -86,7 +86,7 @@ import {
   createBuiltInDependenciesList,
 } from '../../core/es-modules/package-manager/built-in-dependencies-list'
 import { clearAllRegisteredControls } from './canvas-globals'
-import { createEmptySessionStateState } from './interactions/interaction-state'
+import { createEmptySessionStateState } from './canvas-strategies/interaction-state'
 
 // eslint-disable-next-line no-unused-expressions
 typeof process !== 'undefined' &&
