+++ editor/src/templates/editor.tsx
@@ -84,7 +84,7 @@ import { PersistenceMachine } from '../components/editor/persistence/persistence
 import { PersistenceBackend } from '../components/editor/persistence/persistence-backend'
 import { defaultProject } from '../sample-projects/sample-project-utils'
 import { createBuiltInDependenciesList } from '../core/es-modules/package-manager/built-in-dependencies-list'
-import { createEmptySessionStateState } from '../components/canvas/interactions/interaction-state'
+import { createEmptySessionStateState } from '../components/canvas/canvas-strategies/interaction-state'
 
 if (PROBABLY_ELECTRON) {
   let { webFrame } = requireElectron()
